requirements.txt
torch
numpy
opencv-python-headless
setuptools
Pillow
gradio
torchvision
addict
future
lmdb
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
psutil
gradio